DAOS-5157 control: Make GetAttachInfo() backward-compatible
When an older client makes a GetAttachInfo() RPC, it doesn't set
AllRanks: true. This change restores the previous behavior where
only the MS rank info would be returned in order to maintain
compatibility.
